Remove QObject serialization refcounting
The idea here was to keep a count of all references to a QObject from
serialization of another QObject, and use those to keep the object
alive.

This doesn't really make sense, in the end: the important point of
references is that any object usable by the client stays alive, and that
is done with the client ref and grace period. This reference counting
had a lot of bugs and was likely leaking objects.

There are issues with singletons and edge cases here; this is just the
beginning of the changes.