Refactor object initialization and activation
Previously, the idea of QObject initialization was mixed in with the
registration/activation of the object with the connection. That led to
some annoying API and inconsistencies around deactivated objects (when
they are no longer referenced by the client).

Those are now separated, meaning QObjects can be initialized without
using a Connection (which is much more convenient for setting up
signals), and clearing a path for further changes to object management
in Connection.

The Connection.InitObject(Id) functions are replaced with the global
Initialize(). As before, it's never required to initialize manually, but
can be useful to set up signals or other details.

Connection has a list of "active" objects -- objects which the client
has or could take a reference to. Objects are "activated" and added to
this list during serialization, and removed with the same reference
counting / collection process as before.