Replace refGraceTime with sync points
As before, the client can keep any object it might use alive with a
OBJECT_REF, and must OBJECT_DEREF those when it's no longer interested.

Previously, objects with references "in flight" were kept alive until
the client could OBJECT_REF by a timer, and cleaned up by a periodic
scan of objects.

This adds the SYNC message, which the client returns as SYNC_ACK only
after it has handled every message sent before the SYNC. It does not
guarantee that the client hasn't handled messages after the SYNC, but
that isn't relevant here.

Using this, there's a more robust way to keep objects alive. A syncRef
is held for any object that is sent (by reference) to prevent its
removal. Periodically, this set of objects is rotated and SYNC is sent.
Once the client responds with a SYNC_ACK, any object among that set that
doesn't have a clientRef or a new syncRef can be safely removed.